task: package/test pre/post html and typescript plugins
