+++ greenwood.config.js
@@ -4,10 +4,10 @@ const path = require('path');
 
 const META_DESCRIPTION = 'A modern and performant static site generator supporting Web Component based development';
 const FAVICON_HREF = '/assets/favicon.ico';
-const preProcessHTMLTransformPlugin = require('./packages/plugin-preprocess-html/preprocess.transform');
-const postProcessHTMLTransformPlugin = require('./packages/plugin-postprocess-html/postprocess.transform');
+const preProcessHTMLTransformPlugin = require('./packages/plugin-preprocess-html/src/index');
+const postProcessHTMLTransformPlugin = require('./packages/plugin-postprocess-html/src/index');
 const sassTransformPlugin = require('./packages/plugin-sass/src/index');
-const tsTransformPlugin = require('./packages/plugin-typescript/plugin-typescript.transform');
+const tsTransformPlugin = require('./packages/plugin-typescript/src/index');
 
 module.exports = {
   workspace: path.join(__dirname, 'www'),
