+++ packages/cli/package.json
@@ -49,7 +49,6 @@
     "rollup-plugin-multi-input": "^1.1.1",
     "rollup-plugin-postcss": "^3.1.5",
     "rollup-plugin-terser": "^7.0.0",
-    "typescript": "^4.1.3",
     "unified": "^9.2.0"
   },
   "devDependencies": {
