+++ packages/plugin-postprocess-html/package.json
@@ -1,0 +1,31 @@
+{
+  "name": "@greenwood/plugin-postprocess-html",
+  "version": "0.10.0-alpha.0",
+  "description": "A post-transform HTML plugin for Greenwood.",
+  "repository": "https://github.com/ProjectEvergreen/greenwood/tree/master/packages/plugin-postprocess-html",
+  "author": "Grant Hutchinson <grant@hutchdev.ca>",
+  "license": "MIT",
+  "keywords": [
+    "Greenwood",
+    "Web Components",
+    "Lit Element",
+    "Lit Html",
+    "Static Site Generator",
+    "plugin",
+    "postprocess",
+    "html"
+  ],
+  "main": "src/index.js",
+  "files": [
+    "src/"
+  ],
+  "publishConfig": {
+    "access": "public"
+  },
+  "peerDependencies": {
+    "@greenwood/cli": "^0.4.0"
+  },
+  "devDependencies": {
+    "@greenwood/cli": "^0.10.0-alpha.0"
+  }
+}
