+++ packages/plugin-postprocess-html/test/cases/default/greenwood.config.js
@@ -1,0 +1,7 @@
+const postProcessHTMLTransformPlugin = require('../../../src/index');
+
+module.exports = {
+  plugins: [
+    ...postProcessHTMLTransformPlugin()
+  ]
+};
