+++ packages/plugin-preprocess-html/src/index.js
@@ -1,8 +1,8 @@
 const fs = require('fs');
 const path = require('path');
-const TransformInterface = require('../cli/src/transforms/transform.interface');
+const TransformInterface = require('@greenwood/cli/src/transforms/transform.interface');
 
-class TestPostTransform extends TransformInterface {
+class TestTransform extends TransformInterface {
 
   constructor(req, compilation) {
     super(req, compilation, {
@@ -13,21 +13,29 @@ class TestPostTransform extends TransformInterface {
 
   shouldTransform() {
     const { url } = this.request;
+    console.debug(this.request.url);
     const barePath = url.endsWith('/')
       ? `${this.workspace}/pages${url}index`
       : `${this.workspace}/pages${url.replace('.html', '')}`;
       
     return (this.extensions.indexOf(path.extname(url)) >= 0 || path.extname(url) === '') && 
-      (fs.existsSync(`${barePath}.html`) && url.endsWith('transform-example.html'));
+      fs.existsSync(`${barePath}.html`);
   }
 
-  async applyTransform(response) {
+  async applyTransform() {
     
     return new Promise(async (resolve, reject) => {
       try {
-        let body = response.body.replace(/<\/h1>/, `
-        and post process plugin </h1>\n
-      `);
+        let body = `
+        <html>
+          <head>
+            <title>test</title>
+            <meta-outlet></meta-outlet>
+          </head>
+          <body>
+            <h1>test pre process plugin</h1>
+          </body>
+        </html>`;
 
         resolve({
           body,
@@ -44,8 +52,8 @@ class TestPostTransform extends TransformInterface {
 module.exports = () => {
   return [
     {
-      type: 'transform-post',
-      provider: (req, compilation) => new TestPostTransform(req, compilation)
+      type: 'transform-pre',
+      provider: (req, compilation) => new TestTransform(req, compilation)
     }
   ];
 };
