+++ packages/plugin-preprocess-html/test/cases/default/greenwood.config.js
@@ -1,0 +1,7 @@
+const preProcessHTMLTransformPlugin = require('../../../src/index');
+
+module.exports = {
+  plugins: [
+    ...preProcessHTMLTransformPlugin()
+  ]
+};
